Extract polling officer zone menu for elections (votings)
