+++ decidim-elections/lib/decidim/votings/menu.rb
@@ -3,16 +3,6 @@
 module Decidim
   module Votings
     class Menu
-      def self.register_user_menu!
-        Decidim.menu :user_menu do |menu|
-          menu.add_item :decidim_votings_polling_officer_zone,
-                        I18n.t("menu.polling_officer_zone", scope: "decidim.votings.polling_officer_zone"),
-                        decidim.decidim_votings_polling_officer_zone_path,
-                        active: :inclusive,
-                        if: Decidim::Votings::PollingOfficer.polling_officer?(current_user)
-        end
-      end
-
       def self.register_menu!
         Decidim.menu :menu do |menu|
           menu.add_item :votings,
